+++ library/include/ck/library/reference_tensor_operation/cpu/reference_pool_fwd.hpp
