+++ library/src/tensor_operation_instance/gpu/pool_fwd/pool_fwd_instance_common.hpp
